CLDINT-860 improvement of transparent synchronization (Refresh action indicates changes and force next sync); upgrade to Box driver 2.1.x

+++ src/main/java/org/exoplatform/clouddrive/CloudDrive.java
@@ -120,16 +120,16 @@ public abstract class CloudDrive {
     /**
      * Complete work in abstract units.
      * 
-     * @return int
+     * @return long
      */
-    int getComplete();
+    long getComplete();
 
     /**
      * Available work to do in abstract units.
      * 
-     * @return int
+     * @return long
      */
-    int getAvailable();
+    long getAvailable();
   }
 
   /**
